# Improve span creation performance another 3-7%